+++ mympi.global.funcdeclare.h
@@ -76,7 +76,7 @@ extern void mpildsum0(long int*sumptr, int recvid);
 extern void mpifmin(FTYPE*minptr);
 
 extern int set_sizeofmemory(int numbuff, int sizeofdatatype, int numcolumns, int *sizeofmemory);
-extern int set_maxnumsize(int numcolumns, int *maxnumsize);
+extern int set_maxnumsize(int numcolumns, long long int *maxnumsize);
 extern int set_numbuffers(int numcolumns, int *numbuffers);
 extern int gcountmod(int numcolumns);
